+++ src/Farmer/Arm/EventHub.fs
@@ -2,7 +2,6 @@
 module Farmer.Arm.EventHub
 
 open Farmer
-open Farmer.CoreTypes
 open Farmer.EventHub
 
 let namespaces = ResourceType ("Microsoft.EventHub/namespaces", "2017-04-01")
@@ -26,7 +25,7 @@ type Namespace =
             | AutoInflate throughput -> Some throughput
             | ManualInflate -> None)
     interface IArmResource with
-        member this.ResourceName = this.Name
+        member this.ResourceId = namespaces.resourceId this.Name
         member this.JsonModel =
             {| namespaces.Create(this.Name, this.Location, tags = this.Tags) with
                 sku =
@@ -53,7 +52,7 @@ module Namespaces =
           CaptureDestination : CaptureDestination option
           Tags: Map<string,string>  }
         interface IArmResource with
-            member this.ResourceName = this.Name
+            member this.ResourceId = eventHubs.resourceId this.Name
             member this.JsonModel =
                {| eventHubs.Create(this.Name, this.Location, this.Dependencies, this.Tags) with
                       properties =
@@ -68,7 +67,7 @@ module Namespaces =
                                    destination =
                                     {| name = "EventHubArchive.AzureBlockBlob"
                                        properties =
-                                        {| storageAccountResourceId = ResourceId.create(storageAccounts, name).Eval()
+                                        {| storageAccountResourceId = storageAccounts.resourceId(name).Eval()
                                            blobContainer = container
                                         |}
                                     |}
@@ -84,7 +83,7 @@ module Namespaces =
               Location : Location
               Dependencies : ResourceId list }
             interface IArmResource with
-                member this.ResourceName = this.ConsumerGroupName
+                member this.ResourceId = consumerGroups.resourceId (this.EventHub/this.ConsumerGroupName)
                 member this.JsonModel = consumerGroups.Create(this.EventHub/this.ConsumerGroupName, this.Location, this.Dependencies) :> _
 
         type AuthorizationRule =
@@ -93,7 +92,7 @@ module Namespaces =
               Dependencies : ResourceId list
               Rights : AuthorizationRuleRight Set }
             interface IArmResource with
-                member this.ResourceName = this.Name
+                member this.ResourceId = authorizationRules.resourceId this.Name
                 member this.JsonModel =
                     {| authorizationRules.Create(this.Name, this.Location, this.Dependencies) with
                         properties = {| rights = this.Rights |> Set.map string |> Set.toList |}