+++ src/Farmer/Common.fs
@@ -628,12 +628,11 @@ type RoleId =
         match this with
         | RoleId roleId ->
             sprintf "concat('/subscriptions/', subscription().subscriptionId, '/providers/Microsoft.Authorization/roleDefinitions/', '%O')" roleId.Id
-            |> Farmer.CoreTypes.ArmExpression.create
+            |> ArmExpression.create
     member this.Name = match this with (RoleId v) -> v.Name
     member this.Id = match this with (RoleId v) -> v.Id
 
 module Identity =
-    open Farmer.CoreTypes
 
     /// Represents a User Assigned Identity, and the ability to create a Principal Id from it.
     type UserAssignedIdentity =
@@ -649,18 +648,14 @@ module Identity =
 
     type SystemIdentity =
         | SystemIdentity of ResourceId
+        member this.ResourceId = match this with SystemIdentity r -> r
         member private this.CreateExpression field =
-            match this.ResourceId.Type with
-            | None ->
-                failwith "Resource Id must have a type in order to generate a Prinicipal ID"
-            | Some resourceType ->
-                let identity = this.ResourceId.ArmExpression.Value
-                ArmExpression
-                    .create(sprintf "reference(%s, '%s', 'full').identity.%s" identity resourceType.ApiVersion field)
-                    .WithOwner(this.ResourceId)
+            let identity = this.ResourceId.ArmExpression.Value
+            ArmExpression
+                .create(sprintf "reference(%s, '%s', 'full').identity.%s" identity this.ResourceId.Type.ApiVersion field)
+                .WithOwner(this.ResourceId)
         member this.PrincipalId = this.CreateExpression "principalId" |> PrincipalId
         member this.ClientId = this.CreateExpression "clientId"
-        member this.ResourceId = match this with SystemIdentity r -> r
 
     /// Represents an identity that can be assigned to a resource for impersonation.
     type ManagedIdentity =
@@ -1080,7 +1075,7 @@ module Resource =
     /// Creates a unique IArmResource from an arbitrary object.
     let ofObj armObject =
         { new IArmResource with
-             member _.ResourceName = ResourceName (System.Guid.NewGuid().ToString())
+             member _.ResourceId = ResourceId.create (ResourceType("", ""), ResourceName (System.Guid.NewGuid().ToString()))
              member _.JsonModel = armObject }
 
     /// Creates a unique IArmResource from a JSON string containing the output you want.
@@ -1092,5 +1087,5 @@ module Json =
 
 module Subscription =
     /// Gets an ARM expression pointing to the tenant id of the current subscription.
-    let TenantId = Farmer.CoreTypes.ArmExpression.create "subscription().tenantid"
+    let TenantId = ArmExpression.create "subscription().tenantid"