+++ src/Farmer/Types.fs
@@ -1,5 +1,7 @@
 namespace Farmer
 
+open System
+
 /// Represents a name of an ARM resource
 type ResourceName =
     | ResourceName of string
@@ -13,42 +15,14 @@ type ResourceName =
         | r -> r
     member this.Map mapper = match this with ResourceName r -> ResourceName (mapper r)
 
-    static member (+) (a:ResourceName, b:string) = ResourceName(a.Value + "/" + b)
-    static member (+) (a:ResourceName, b:ResourceName) = a + b.Value
-    static member (/) (a:ResourceName, b:string) = a + b
-    static member (/) (a:ResourceName, b:ResourceName) = a + b.Value
+    static member (/) (a:ResourceName, b:string) = ResourceName(a.Value + "/" + b)
+    static member (/) (a:ResourceName, b:ResourceName) = a / b.Value
+    static member (-) (a:ResourceName, b:string) = ResourceName(a.Value + "-" + b)
+    static member (-) (a:ResourceName, b:ResourceName) = a - b.Value
 
 type Location =
     | Location of string
     member this.ArmValue = match this with Location location -> location.ToLower()
-
-/// An Azure ARM resource value which can be mapped into an ARM template.
-type IArmResource =
-    /// The name of the resource, to uniquely identify against other resources in the template.
-    abstract member ResourceName : ResourceName
-    /// A raw object that is ready for serialization directly to JSON.
-    abstract member JsonModel : obj
-
-/// Represents a high-level configuration that can create a set of ARM Resources.
-type IBuilder =
-    /// Given a location and the currently-built resources, returns a set of resource actions.
-    abstract member BuildResources : Location -> IArmResource list
-    /// Provides the resource name that other resources should use when depending upon this builder.
-    abstract member DependencyName : ResourceName
-
-/// Represents a high-level configuration that can create a set of ARM Resources at the subscription scope.
-type ISubscriptionResourceBuilder =
-    // Gets the outputs specified by this resource
-    abstract member Outputs: Map<string,string>
-    /// Given a location and the currently-built resources, returns a set of resource actions.
-    abstract member BuildResources : string -> {|Resources: IArmResource list; DeploymentName: string|}
-    /// Gets the list of tasks to execute after a deployment
-    abstract member RunPostDeployTasks : unit -> Result<string,string> list
-
-namespace Farmer.CoreTypes
-
-open Farmer
-open System
 
 type ResourceType =
     | ResourceType of path:string * version:string
@@ -57,21 +31,43 @@ type ResourceType =
     member this.ApiVersion = match this with ResourceType (_, v) -> v
 
 type ResourceId =
-    { Type : ResourceType option
+    { Type : ResourceType
       ResourceGroup : string option
       Name : ResourceName
       Segments : ResourceName list }
-    static member Empty = { Type = None; ResourceGroup = None; Name = ResourceName.Empty; Segments = [] }
-    member this.WithType resourceType = { this with Type = Some resourceType }
-    static member create (name:ResourceName, ?group) =
-        { ResourceId.Empty with Name = name; ResourceGroup = group }
-    static member create (name:string, ?group) =
-        ResourceId.create (ResourceName name, ?group = group)
     static member create (resourceType:ResourceType, name:ResourceName, ?group:string) =
-        { ResourceId.Empty with Type = Some resourceType; ResourceGroup = group; Name = name }
+        { Type = resourceType; ResourceGroup = group; Name = name; Segments = [] }
     static member create (resourceType:ResourceType, name:ResourceName, [<ParamArray>] resourceSegments:ResourceName []) =
-        { ResourceId.Empty with Type = Some resourceType; Name = name; Segments = List.ofArray resourceSegments }
+        { Type = resourceType; Name = name; ResourceGroup = None; Segments = List.ofArray resourceSegments }
 
+type ResourceType with
+    member this.resourceId name = ResourceId.create (this, name)
+    member this.resourceId name = this.resourceId (ResourceName name)
+    member this.resourceId (name, [<ParamArray>] resourceSegments:ResourceName []) = ResourceId.create (this, name, resourceSegments)
+
+/// An Azure ARM resource value which can be mapped into an ARM template.
+type IArmResource =
+    /// The name of the resource, to uniquely identify against other resources in the template.
+    abstract member ResourceId : ResourceId
+    /// A raw object that is ready for serialization directly to JSON.
+    abstract member JsonModel : obj
+
+/// Represents a high-level configuration that can create a set of ARM Resources.
+type IBuilder =
+    /// Given a location and the currently-built resources, returns a set of resource actions.
+    abstract member BuildResources : Location -> IArmResource list
+    /// Provides the ResourceId that other resources should use when depending upon this builder.
+    abstract member ResourceId : ResourceId
+    
+/// Represents a high-level configuration that can create a set of ARM Resources at the subscription scope.
+type ISubscriptionResourceBuilder =
+    // Gets the outputs specified by this resource
+    abstract member Outputs: Map<string,string>
+    /// Given a location and the currently-built resources, returns a set of resource actions.
+    abstract member BuildResources : deploymentNameSuffix: string option -> {| Resources: IArmResource list; DeploymentName: string|}
+    /// Gets the list of tasks to execute after a deployment
+    abstract member RunPostDeployTasks : unit -> Result<string,string> list
+    
 /// Represents an expression used within an ARM template
 type ArmExpression =
     private | ArmExpression of expression:string * owner:ResourceId option
@@ -94,8 +90,8 @@ type ArmExpression =
         | None -> sprintf "[%s]" this.Value
     /// Sets the owning resource on this ARM Expression.
     member this.WithOwner(owner:ResourceId) = match this with ArmExpression (e, _) -> ArmExpression(e, Some owner)
-    /// Sets the owning resource on this ARM Expression.
-    member this.WithOwner(owner:ResourceName) = this.WithOwner(ResourceId.create owner)
+    // /// Sets the owning resource on this ARM Expression.
+    // member this.WithOwner(owner:ResourceName) = this.WithOwner(ResourceId.create owner)
 
     /// Evaluates the expression for emitting into an ARM template. That is, wraps it in [].
     static member Eval (expression:ArmExpression) = expression.Eval()
@@ -114,18 +110,20 @@ type ArmExpression =
 
 type ResourceId with
     member this.ArmExpression =
-        match this with
-        | { Type = None } ->
-            this.Name.Value |> sprintf "string('%s')" |> ArmExpression.create
-        | { Type = Some resourceType } ->
-            [ match this.ResourceGroup with Some rg -> rg | None -> ()
-              resourceType.Type
+        match this.Type.Type with
+        | "" ->
+            sprintf "string('%s')" this.Name.Value
+            |> ArmExpression.create
+        | _ ->
+            [ yield! Option.toList this.ResourceGroup
+              this.Type.Type
               this.Name.Value
               for segment in this.Segments do segment.Value ]
             |> List.map (sprintf "'%s'")
             |> String.concat ", "
             |> sprintf "resourceId(%s)"
             |> ArmExpression.create
+
     /// Evaluates the expression for emitting into an ARM template. That is, wraps it in [].
     member this.Eval() = this.ArmExpression.Eval()
 
@@ -135,7 +133,7 @@ type ArmExpression with
                      .WithOwner(resourceId)
 
 type ResourceType with
-    member this.Create(name:ResourceName, ?location:Location, ?dependsOn:ResourceId list, ?tags:Map<string,string>) =
+    member this.Create(name:ResourceName, ?location:Location, ?dependsOn:ResourceId seq, ?tags:Map<string,string>) =
         match this with
         | ResourceType (path, version) ->
             {| ``type`` = path
@@ -144,7 +142,7 @@ type ResourceType with
                location = location |> Option.map(fun r -> r.ArmValue) |> Option.toObj
                dependsOn =
                 dependsOn
-                |> Option.map (List.map(fun r -> r.Eval()) >> box)
+                |> Option.map (Seq.map(fun r -> r.Eval()) >> Seq.toArray >> box)
                 |> Option.toObj
                tags = tags |> Option.map box |> Option.toObj |}
 
@@ -170,10 +168,10 @@ type Builder = Location -> IArmResource list
 /// A resource that will automatically be created by Farmer.
 type AutoCreationKind<'TConfig> =
     /// A resource that will automatically be created by Farmer with an explicit (user-defined) name.
-    | Named of ResourceName
+    | Named of ResourceId
     /// A resource that will automatically be created by Farmer with a name that is derived based on the configuration.
-    | Derived of ('TConfig -> ResourceName)
-    member this.CreateResourceName config =
+    | Derived of ('TConfig -> ResourceId)
+    member this.resourceId config =
         match this with
         | Named r -> r
         | Derived f -> f config
@@ -181,7 +179,7 @@ type AutoCreationKind<'TConfig> =
 /// A related resource that is created externally to this Farmer resource.
 type ExternalKind =
     /// The name of the resource that will be created by Farmer, but is explicitly linked by the user.
-    | Managed of ResourceName
+    | Managed of ResourceId
     /// A Resource Id that is created externally from Farmer and already exists in Azure.
     | Unmanaged of ResourceId
 
@@ -189,32 +187,36 @@ type ExternalKind =
 type ResourceRef<'TConfig> =
     | AutoCreate of AutoCreationKind<'TConfig>
     | External of ExternalKind
-    member this.CreateResourceId config =
+    member this.resourceId config =
         match this with
-        | External (Managed r) -> ResourceId.create r
+        | External (Managed r) -> r
         | External (Unmanaged r) -> r
-        | AutoCreate r -> r.CreateResourceName config |> ResourceId.create
+        | AutoCreate r -> r.resourceId config
 
 [<AutoOpen>]
 module ResourceRef =
     /// Creates a ResourceRef which is automatically created and derived from the supplied config.
     let derived derivation = derivation |> Derived |> AutoCreate
+    let named (resourceType:ResourceType) name = AutoCreate(Named(ResourceId.create(resourceType, name)))
+    let managed (resourceType:ResourceType) name = External(Managed(ResourceId.create(resourceType, name)))
     /// An active pattern that returns the resource name if the resource should be set as a dependency.
     /// In other words, all cases except External Unmanaged.
     let (|DependableResource|_|) config = function
         | External (Managed r) -> Some (DependableResource r)
-        | AutoCreate r -> Some(DependableResource(r.CreateResourceName config))
+        | AutoCreate r -> Some(DependableResource(r.resourceId config))
         | External (Unmanaged _) -> None
     /// An active pattern that returns the resource name if the resource should be deployed. In other
     /// words, AutoCreate only.
     let (|DeployableResource|_|) config = function
-        | AutoCreate c -> Some (DeployableResource(c.CreateResourceName config))
+        | AutoCreate c -> Some (DeployableResource(c.resourceId config))
         | External _ -> None
     /// An active pattern that returns the resource name if the resource if external.
     let (|ExternalResource|_|) = function
-        | AutoCreate c -> None
-        | External (Managed r) -> Some (ResourceId.create r)
-        | External (Unmanaged r) -> Some r
+        | AutoCreate _ ->
+            None
+        | External (Managed r)
+        | External (Unmanaged r) ->
+            Some r
 
 /// Whether a specific feature is active or not.
 type FeatureFlag =
@@ -262,14 +264,17 @@ type Deployment =
       Template : ArmTemplate
       PostDeployTasks : (unit -> Result<string ,string> list) list}
     interface IDeploymentBuilder with
-        member this.BuildDeployment _ = this
+        member this.BuildDeployment _ _ = this
 
 and IDeploymentBuilder =
-    abstract member BuildDeployment : string -> Deployment
+    abstract member BuildDeployment : string -> string option -> Deployment
 
 module Deployment =
-    let build defaultName (builder:#IDeploymentBuilder) = builder.BuildDeployment defaultName
+    let uniqueSuffix () = Some <| System.DateTime.UtcNow.ToString "yyyyMMddTHHmm" // ISO 8601 Date Time format to minute resolution
+    let buildWithSuffix defaultName deploymentSuffix (builder:#IDeploymentBuilder) = builder.BuildDeployment defaultName deploymentSuffix
+    let build defaultName (builder:#IDeploymentBuilder) = builder.BuildDeployment defaultName (uniqueSuffix ())
     let getTemplate defaultName (builder:#IDeploymentBuilder) = (builder |> build defaultName).Template
+    let getTemplateWithSuffix defaultName deploymentSuffix (builder:#IDeploymentBuilder) = (builder |> buildWithSuffix defaultName deploymentSuffix).Template
 
 module internal DeterministicGuid =
     open System